refactor: extract shared window iterators from bootstrap and null model
The bootstrap (compute_divergence_bootstrap.py) and null model
(compute_null_model.py) scripts had nearly identical data-loading
and window-iteration code for both semantic and lexical channels
(~120 lines duplicated across 4 functions).

Extract three shared helpers into _divergence_io.py:
- iter_semantic_windows(): yields (year, window, X, Y, extra_rng)
- iter_lexical_windows(): yields (year, window, texts, texts, extra_rng)
- fit_lexical_vectorizer(): fits TF-IDF on full corpus
- _make_window_rngs(): moved from compute_null_model (re-exported)

Each script now has thin channel drivers (~6 lines each) that
compose the iterator with their core operation (permutation test
or bootstrap resampling) via a shared collector function.

Also removes dead _nan_row() from compute_null_model.py (iterators
skip invalid windows, no NaN rows emitted).
